# packages/backend/src/infrastructure/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Equipment Rental System API
  description: |
    A comprehensive API for managing equipment rentals, reservations, and members.
    Built using hexagonal architecture with TypeScript.

    ## Features
    - Equipment management and availability tracking
    - Member registration and tier-based benefits
    - Rental creation, extension, and return workflows
    - Reservation system with payment authorization
    - Damage assessment and late fee calculation
    - Payment processing integration
    - Notification system

    ## Architecture
    This API follows hexagonal (ports and adapters) architecture principles,
    ensuring clean separation of concerns and testability.
  version: 1.0.0
  contact:
    name: Equipment Rental System
    email: support@equipmentrental.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: http://localhost:3000/api
    description: API base path
...
paths:
...
  /api/members/{memberId}/rentals:
    get:
      summary: Get member rentals
      description: Retrieves all rentals for a specific member
      tags:
        - Members
      parameters:
        - $ref: '#/components/parameters/MemberId'
      responses:
        '200':
          description: List of member rentals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberRental'
              example:
                - rentalId: rent_001
                  equipmentId: eq_123
                  equipmentName: Professional Camera
                  startDate: '2024-01-10T00:00:00.000Z'
                  endDate: '2024-01-15T00:00:00.000Z'
                  status: ACTIVE
                  totalCost: 250.00
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/rentals:
    post:
      summary: Create a new rental
      description: Creates a new rental for a member with payment processing
      tags:
        - Rentals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRentalRequest'
            example:
              equipmentId: eq_123
              memberId: mem_789
              startDate: '2024-01-15T00:00:00.000Z'
              endDate: '2024-01-20T00:00:00.000Z'
              paymentMethod:
                type: credit_card
                token: tok_visa_4242
              notificationChannel: EMAIL
      responses:
        '201':
          description: Rental created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRentalResponse'
              example:
                rentalId: rent_001
                totalCost: 225.00
                discountApplied: 25.00
                transactionId: txn_123456
                paymentStatus: succeeded
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/rentals/{rentalId}:
    get:
      summary: Get rental by ID
      description: Retrieves detailed information about a specific rental
      tags:
        - Rentals
      parameters:
        - $ref: '#/components/parameters/RentalId'
      responses:
        '200':
          description: Rental details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rental'
              example:
                rentalId: rent_001
                equipmentId: eq_123
                memberId: mem_789
                startDate: '2024-01-15T00:00:00.000Z'
                endDate: '2024-01-20T00:00:00.000Z'
                status: ACTIVE
                totalCost: 225.00
                conditionAtStart: EXCELLENT
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/rentals/{rentalId}/return:
    put:
      summary: Return a rental
      description: Processes the return of rented equipment with condition assessment
      tags:
        - Rentals
      parameters:
        - $ref: '#/components/parameters/RentalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnRentalRequest'
            example:
              conditionAtReturn: GOOD
              paymentMethod:
                type: credit_card
                token: tok_visa_4242
              notificationChannel: EMAIL
      responses:
        '200':
          description: Rental returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnRentalResponse'
              example:
                rentalId: rent_001
                totalCost: 225.00
                lateFee: 0.00
                damageFee: 0.00
                transactionId: txn_789012
                paymentStatus: succeeded
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/rentals/{rentalId}/extend:
    put:
      summary: Extend a rental
      description: Extends the rental period for additional days with payment
      tags:
        - Rentals
      parameters:
        - $ref: '#/components/parameters/RentalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtendRentalRequest'
            example:
              additionalDays: 3
              paymentMethod:
                type: credit_card
                token: tok_visa_4242
              notificationChannel: EMAIL
      responses:
        '200':
          description: Rental extended successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtendRentalResponse'
              example:
                rentalId: rent_001
                newEndDate: '2024-01-23T00:00:00.000Z'
                additionalCost: 135.00
                transactionId: txn_345678
                paymentStatus: succeeded
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/rentals/status/overdue:
    get:
      summary: Get overdue rentals
      description: Retrieves all rentals that are past their due date
      tags:
        - Rentals
      responses:
        '200':
          description: List of overdue rentals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rental'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/rentals/member/{memberId}:
    get:
      summary: Get rentals by member
      description: Retrieves all rentals for a specific member (alias endpoint)
      tags:
        - Rentals
      parameters:
        - $ref: '#/components/parameters/MemberId'
      responses:
        '200':
          description: List of member rentals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberRental'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
...
    MemberId:
      name: memberId
      in: path
      required: true
      description: Unique identifier for the member
      schema:
        type: string
        example: mem_789
    RentalId:
      name: rentalId
      in: path
      required: true
      description: Unique identifier for the rental
      schema:
        type: string
        example: rent_001
...
  schemas:
...
    MemberRental:
      type: object
      required:
        - rentalId
        - equipmentId
        - equipmentName
        - startDate
        - endDate
        - status
        - totalCost
      properties:
        rentalId:
          type: string
          description: Unique identifier for the rental
        equipmentId:
          type: string
          description: ID of rented equipment
        equipmentName:
          type: string
          description: Name of rented equipment
        startDate:
          type: string
          format: date-time
          description: Rental start date
        endDate:
          type: string
          format: date-time
          description: Rental end date
        status:
          type: string
          enum: [ACTIVE, RETURNED, OVERDUE, CANCELLED]
          description: Current rental status
        totalCost:
          type: number
          format: float
          description: Total cost of rental

    # Rental Schemas
    CreateRentalRequest:
      type: object
      required:
        - equipmentId
        - memberId
        - startDate
        - endDate
        - paymentMethod
      properties:
        equipmentId:
          type: string
          description: ID of equipment to rent
        memberId:
          type: string
          description: ID of member creating rental
        startDate:
          type: string
          format: date-time
          description: Rental start date (ISO 8601)
        endDate:
          type: string
          format: date-time
          description: Rental end date (ISO 8601)
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        notificationChannel:
          type: string
          enum: [EMAIL, SMS]
          description: Preferred notification channel
    CreateRentalResponse:
      type: object
      required:
        - rentalId
        - totalCost
        - discountApplied
        - transactionId
        - paymentStatus
      properties:
        rentalId:
          type: string
          description: Unique identifier for the created rental
        totalCost:
          type: number
          format: float
          description: Total cost after discounts
        discountApplied:
          type: number
          format: float
          description: Amount of discount applied
        transactionId:
          type: string
          description: Payment transaction ID
        paymentStatus:
          type: string
          enum: [succeeded, pending, failed]
          description: Status of payment
    ReturnRentalRequest:
      type: object
      required:
        - conditionAtReturn
      properties:
        conditionAtReturn:
          type: string
          enum: [EXCELLENT, GOOD, FAIR, POOR, NEEDS_MAINTENANCE]
          description: Condition of equipment upon return
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        notificationChannel:
          type: string
          enum: [EMAIL, SMS]
          description: Preferred notification channel
    ReturnRentalResponse:
      type: object
      required:
        - rentalId
        - totalCost
        - lateFee
        - damageFee
      properties:
        rentalId:
          type: string
          description: ID of returned rental
        totalCost:
          type: number
          format: float
          description: Total cost including fees
        lateFee:
          type: number
          format: float
          description: Late fee charged
        damageFee:
          type: number
          format: float
          description: Damage fee charged
        transactionId:
          type: string
          description: Payment transaction ID for fees
        paymentStatus:
          type: string
          enum: [succeeded, pending, failed]
          description: Status of fee payment
        refundTransactionId:
          type: string
          description: Refund transaction ID if applicable
        refundAmount:
          type: number
          format: float
          description: Amount refunded if applicable
    ExtendRentalRequest:
      type: object
      required:
        - additionalDays
        - paymentMethod
      properties:
        additionalDays:
          type: number
          description: Number of days to extend
          minimum: 1
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        notificationChannel:
          type: string
          enum: [EMAIL, SMS]
          description: Preferred notification channel
    ExtendRentalResponse:
      type: object
      required:
        - rentalId
        - newEndDate
        - additionalCost
        - transactionId
        - paymentStatus
      properties:
        rentalId:
          type: string
          description: ID of extended rental
        newEndDate:
          type: string
          format: date-time
          description: New end date after extension
        additionalCost:
          type: number
          format: float
          description: Cost of extension
        transactionId:
          type: string
          description: Payment transaction ID
        paymentStatus:
          type: string
          enum: [succeeded, pending, failed]
          description: Status of payment
    Rental:
      type: object
      required:
        - rentalId
        - equipmentId
        - memberId
        - startDate
        - endDate
        - status
        - totalCost
        - conditionAtStart
      properties:
        rentalId:
          type: string
          description: Unique identifier for the rental
        equipmentId:
          type: string
          description: ID of rented equipment
        memberId:
          type: string
          description: ID of member who rented
        startDate:
          type: string
          format: date-time
          description: Rental start date
        endDate:
          type: string
          format: date-time
          description: Rental end date
        status:
          type: string
          enum: [ACTIVE, RETURNED, OVERDUE, CANCELLED]
          description: Current rental status
        totalCost:
          type: number
          format: float
          description: Total cost of rental
        conditionAtStart:
          type: string
          enum: [EXCELLENT, GOOD, FAIR, POOR]
          description: Equipment condition at rental start
        conditionAtReturn:
          type: string
          enum: [EXCELLENT, GOOD, FAIR, POOR, NEEDS_MAINTENANCE]
          description: Equipment condition at return
        actualReturnDate:
          type: string
          format: date-time
          description: Actual date equipment was returned
        lateFee:
          type: number
          format: float
          description: Late fee if applicable

    # Reservation Schemas
...
    PaymentMethod:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          description: Type of payment method
          enum: [credit_card, debit_card, cash]
        token:
          type: string
          description: Payment token from payment provider
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
              additionalProperties: true
    ValidationErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR]
              description: Validation error code
            message:
              type: string
              description: Human-readable error message
            fields:
              type: array
              description: List of field validation errors
              items:
                type: object
                required:
                  - field
                  - message
                properties:
                  field:
                    type: string
                    description: Field name with error
                  message:
                    type: string
                    description: Validation error message
  responses:
    BadRequest:
      description: Invalid request parameters or validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: Invalid request parameters
              fields:
                - field: startDate
                  message: Start date must be in the future
                - field: endDate
                  message: End date must be after start date
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            equipmentNotFound:
              value:
                error:
                  code: EQUIPMENT_NOT_FOUND
                  message: Equipment not found
                  details:
                    equipmentId: eq_123
            memberNotFound:
              value:
                error:
                  code: MEMBER_NOT_FOUND
                  message: Member not found
                  details:
                    memberId: mem_789
            rentalNotFound:
              value:
                error:
                  code: RENTAL_NOT_FOUND
                  message: Rental not found
                  details:
                    rentalId: rent_001
    Conflict:
      description: Resource conflict or business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            equipmentNotAvailable:
              value:
                error:
                  code: EQUIPMENT_NOT_AVAILABLE
                  message: Equipment is not available for the requested period
                  details:
                    equipmentId: eq_123
                    startDate: '2024-01-15'
                    endDate: '2024-01-20'
            rentalAlreadyReturned:
              value:
                error:
                  code: RENTAL_ALREADY_RETURNED
                  message: Rental has already been returned
                  details:
                    rentalId: rent_001
    Forbidden:
      description: Operation not allowed due to business rules
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            memberSuspended:
              value:
                error:
                  code: MEMBER_SUSPENDED
                  message: Member account is suspended
                  details:
                    memberId: mem_789
            rentalLimitExceeded:
              value:
                error:
                  code: RENTAL_LIMIT_EXCEEDED
                  message: Member has reached maximum number of active rentals
                  details:
                    memberId: mem_789
                    limit: 5
                    current: 5
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_SERVER_ERROR
              message: An unexpected error occurred